+++ addon/locales/en/components/layers-dialogs/settings.js
@@ -44,6 +44,9 @@ export default {
   'tile': {
     'url-textbox': {
       'caption': 'Url'
+    },
+    'no-wrap-checkbox': {
+      'caption': 'Layer\'s noWrap property'
     }
   },