+++ addon/locales/en/layers-styles/gradient-tools.js
@@ -1,0 +1,4 @@
+export default {
+  'fill-gradient': 'Apply fill Gradient',
+  'stroke-gradient': 'Apply stroke Gradient'
+};