+++ addon/locales/ru/components/gradients/gradient-edit.js
@@ -1,0 +1,5 @@
+export default {
+  'gradientName': 'Name of the gradient',
+  'gradientColorStart': 'Initial gradient color',
+  'gradientColorEnd': 'The final color of the gradient'
+};