+++ addon/locales/ru/components/layers-dialogs/copy.js
@@ -1,4 +1,10 @@
 export default {
   'caption': 'Создать копию слоя',
+  'approve-button': {
+    'caption': 'Ok'
+  },
+  'deny-button': {
+    'caption': 'Отмена'
+  },
   'layer-name-postfix': '(Копия)'
 };