+++ addon/locales/ru/translations.js
@@ -5,6 +5,7 @@ import LayersStyles from './layers-styles';
 import MarkersStyles from './markers-styles';
 import MapCommands from './map-commands';
 import MapTools from './map-tools';
+import Gradients from './gradients';
 
 export default {
   'models': Models,
@@ -13,5 +14,6 @@ export default {
   'layers-styles': LayersStyles,
   'markers-styles': MarkersStyles,
   'map-commands': MapCommands,
-  'map-tools': MapTools
+  'map-tools': MapTools,
+  'gradients': Gradients
 };