+++ scripts/deploy-to-gh-pages.sh
@@ -6,23 +6,35 @@ set -e
 # Define repository relative GitHub address.
 repositoryRelativeGitHubAddress="Flexberry/ember-flexberry-gis"
 
+# Define branch name with postfix.
+fullBranchName="${TRAVIS_BRANCH}"
+if [ ! -z "$1" ]
+then
+  fullBranchName+="$1"
+fi
+
 # Clone project into 'repository' subfolder && move to it.
 echo "Prepare for deploy to gh-pages."
+if [ -d repository ]
+then
+  rm -rf "repository"
+fi
+
 echo "Clone ${repositoryRelativeGitHubAddress} repository & checkout latest version of gh-pages branch."
 git clone --depth=50 --branch=gh-pages "https://github.com/${repositoryRelativeGitHubAddress}.git" repository
 cd repository
 
 # Remove results of previous deploy (for current branch) & recreate directory.
-echo "Remove results of previous deploy (for ${TRAVIS_BRANCH} branch)."
-rm -rf "${TRAVIS_BRANCH}"
-mkdir "${TRAVIS_BRANCH}"
+echo "Remove results of previous deploy (for ${fullBranchName} branch)."
+rm -rf "${fullBranchName}"
+mkdir "${fullBranchName}"
 
-# Copy builded ember application from 'dist' folder into 'repository/${TRAVIS_BRANCH}'.
-echo "Copy builded ember application (for ${TRAVIS_BRANCH} branch)."
-cp -r ../dist/* "${TRAVIS_BRANCH}"
+# Copy builded ember application from 'dist' folder into 'repository/${fullBranchName}'.
+echo "Copy builded ember application (for ${fullBranchName} branch)."
+cp -r ../dist/* "${fullBranchName}"
 
-# Autodocumentation for 'master' and 'develop' branches.
-if [ ${TRAVIS_BRANCH} == "master" ] || [ ${TRAVIS_BRANCH} == "develop" ]
+# Autodocumentation for 'master' and 'develop' branches (if second argument isn't "no-doc").
+if [[ ${TRAVIS_BRANCH} == "master" || ${TRAVIS_BRANCH} == "develop" ]] && [[ -z "$2" || ${2} != "no-doc" ]]
 then
   cd ..
 
@@ -54,7 +66,7 @@ git config user.email "mail@flexberry.net"
 
 echo "Commit & push changes."
 git add --all
-git commit -m "Update gh-pages for ${TRAVIS_BRANCH} branch"
+git commit -m "Update gh-pages for ${fullBranchName} branch"
 
 # Redirect any output to /dev/null to hide any sensitive credential data that might otherwise be exposed.
 git push --force --quiet "https://${GH_TOKEN}@github.com/${repositoryRelativeGitHubAddress}.git" > /dev/null 2>&1